Added AVX2 support for BF16 kernels: Row major
 - Currently the BF16 kernels uses the AVX512 VNNI instructions.
   In order to support AVX2 kernels, the BF16 input has to be converted
   to F32 and then the F32 kernels has to be executed.
 - Added un-pack function for the B-Matrix, which does the unpacking of
   the Re-ordered BF16 B-Matrix and converts it to Float.
 - Added a kernel, to convert the matrix data from Bf16 to F32 for the
   give input.
 - Added a new path to the BF16 5LOOP to work with the BF16 data, where
   the packed/unpacked A matrix is converted from BF16 to F32. The
   packed B matrix is converted from BF16 to F32 and the re-ordered B
   matrix is unre-ordered and converted to F32 before feeding to the
   F32 micro kernels.
 - Removed AVX512 condition checks in BF16 code path.
 - Added the Re-order reference code path to support BF16 AVX2.
 - Currently the F32 AVX-2 kernels supports only F32 BIAS support.
   Added BF16 support for BIAS post-op in F32 AVX2 kernels.
 - Bug fix in the test input generation script.

AMD Internal : [SWLCSG - 3281]

+++ bench/bench_aocl_gemm/bench_lpgemm_helpers.h
@@ -261,7 +261,7 @@ static inline void fill_array_bfloat16( void* arr, dim_t size )
 #endif
     for ( dim_t i = 0; i < size; ++i )
     {
-        c_float[i] = (i % 5 );
+        c_float[i] = (float)(i % 5);
     }
     convert_float_arr_to_bf16( c_float, arr, size );
     if ( c_float != NULL )
